refactor: resolve #1441 — move Orientation + ASHRAE-140 leaf types to fluxion-core
Breaks the sim ↔ validation cycle by hoisting 13 pure-data ASHRAE-140 leaf
types (Orientation, WindowArea, ConstructionType, ShadingType, ShadingDevice,
GlassType, WindowSpec, InternalLoads, HvacSchedule, NightVentilation,
BuildingType, GeometrySpec, ConductanceReferences) from
fluxion::validation::ashrae_140_cases into fluxion_core::ashrae_cases.

The non-leaf types (ASHRAE140Case, CaseSpec, CaseBuilder, CommonWall,
ConstructionSpec) stay in validation because they have upward deps to
crate::sim::construction, crate::sim::hvac, and crate::physics::constants.

5 direct + 3 indirect sim callers updated:
- src/sim/solar.rs            — Orientation, WindowArea
- src/sim/construction.rs     — Orientation
- src/sim/per_surface_conduction.rs — Orientation
- src/sim/invariant_checker.rs — Orientation
- src/sim/shading.rs          — WindowArea, Orientation
- src/sim/thermal_model_core.rs   — Orientation, ShadingType (CaseSpec stays)
- src/sim/thermal_model_data.rs   — NightVentilation, Orientation
- src/sim/thermal_model_iterative.rs — GeometrySpec, Orientation, WindowArea
- The legacy
  in sim/solar.rs is deleted per the issue body.

The validation module re-exports each moved type at its original path so
existing  (and friends)
paths continue to resolve. cargo check --workspace passes; the only build
failure is the pre-existing  (incidence_deg
field — broken at HEAD, unrelated to this refactor).

ARCHITECTURE.md §Workspace Layout updated:
- §'Cycle break (#1441)' added (5-cycle-marker table, re-export shim pattern,
  regression guard description).
- §'Remaining cycles' shrinks — the sim ↔ validation bullet is removed; the
  physics ↔ sim/continuous cycle is the next target.

scripts/check_ashrae_cases_cycle.py: regression guard that enforces three
invariants — fluxion-core has no upward deps, src/sim has no Orientation
import from validation, fluxion_core::ashrae_cases contains all 13 leaf
types.
